feat(1/3): move llm to a separate sublibrary
