+++ walledeval/llm/__init__.py
@@ -1,0 +1,11 @@
+# walledeval/llm/__init__.py
+
+from walledeval.llm.core import LLM
+from walledeval.llm.huggingface import hf_models, HF_LLM
+from walledeval.llm.claude import Claude
+
+__all__ = [
+    "LLM",
+    "hf_models", "HF_LLM",
+    "Claude"
+]
